fix: add default path to nav to login page; handle user fullName on user sign-up and send along to supabase on registration; add sign-out connectivity to sidebar; remove duplicative user info
