+++ src/App.tsx
@@ -21,6 +21,7 @@ export const App: React.FC = () => {
     <AuthContextProvider>
       <Router>
         <Routes>
+          <Route path="/" element={<Navigate to="/login" />} />
           <Route path="/login" element={<Login />} />
           <Route path="/register" element={<Register />} />
           <Route path="/forgot-password" element={<ForgotPassword />} />
